+++ lume-bar.js
@@ -21,6 +21,7 @@ class State {
   constructor() {
     const restore = sessionStorage.getItem(this.key);
     this.state = restore ? JSON.parse(restore) : {};
+    this.state.open = localStorage.getItem(this.key) === "open";
   }
 
   set(key, value) {
@@ -39,6 +40,7 @@ class State {
 
   save() {
     sessionStorage.setItem(this.key, JSON.stringify(this.state));
+    localStorage.setItem(this.key, this.state.open ? "open" : "closed");
   }
 }