+++ CHANGELOG.md
@@ -6,9 +6,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.1.7] - Unreleased
+## [0.1.7] - 2025-06-11
 ### Added
 - Allow to place icons in the item title.
+- Save in localStorage if th bar was visible the last time.
 
 ### Changed
 - Show automatically the first tab on open.
@@ -62,7 +63,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2025-05-06
 First version
 
-[0.1.7]: https://github.com/lumeland/bar/compare/v0.1.6...HEAD
+[0.1.7]: https://github.com/lumeland/bar/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/lumeland/bar/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/lumeland/bar/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/lumeland/bar/compare/v0.1.3...v0.1.4